Move userAgent substring generation from Bridge to Strada namespace
