+++ strada/src/main/kotlin/dev/hotwire/strada/Bridge.kt
@@ -125,11 +125,6 @@ class Bridge internal constructor(webView: WebView) {
             }
         }
 
-        fun userAgentSubstring(componentFactories: List<BridgeComponentFactory<*,*>>): String {
-            val components = componentFactories.joinToString(" ") { it.name }
-            return "bridge-components: [$components]"
-        }
-
         @VisibleForTesting
         internal fun initialize(bridge: Bridge) {
             instances.add(bridge)
