+++ strada/src/test/kotlin/dev/hotwire/strada/UserAgentTest.kt
@@ -1,0 +1,17 @@
+package dev.hotwire.strada
+
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+class UserAgentTest {
+    @Test
+    fun userAgentSubstring() {
+        val factories = listOf(
+            BridgeComponentFactory("one", TestData::OneBridgeComponent),
+            BridgeComponentFactory("two", TestData::TwoBridgeComponent)
+        )
+
+        val userAgentSubstring = Strada.userAgentSubstring(factories)
+        assertEquals(userAgentSubstring, "bridge-components: [one two]")
+    }
+}
